📦 feat(route.ts): add PATCH route to update server invite code
This commit adds a new PATCH route to update the invite code of a server. The route receives a serverId parameter and checks if the current user is authorized. If the user is not authorized, a 401 Unauthorized response is returned. If the serverId is missing, a 400 Bad Request response is returned. Otherwise, the server's invite code is updated with a new UUID and the updated server object is returned in the response body. In case of any errors, a 500 Internal Server Error response is returned.