📦 feat(route.ts): add DELETE route for deleting a channel in a server
This commit adds a new file `route.ts` which contains the implementation of the DELETE route for deleting a channel in a server. The route expects the `serverId` and `channelId` as parameters in the request URL.

The implementation first checks if the user is authorized by retrieving the current profile. If the user is not authorized, a 401 Unauthorized response is returned. Then, it checks if the `serverId` and `channelId` parameters are present in the request. If not, a 400 Bad Request response is returned.

Next, it uses the Prisma ORM to update the server by deleting the specified channel. The update operation is performed only if the user has the role of ADMIN or MODERATOR in the server. The channel with the specified `channelId` is deleted, but only if its name is not 'general'.

If the update operation is successful, a 200 OK response is returned with the updated server object in the response body. If any error occurs during the process, a 500 Internal Server Error response is returned.

This route is intended to be used for deleting channels in a server.