📦 chore(current-profile-pages.ts): add getCurrentProfile function to retrieve the current user's profile from the database
The getCurrentProfile function is added to the current-profile-pages.ts file. This function uses the getAuth function from the @clerk/nextjs/server package to retrieve the userId from the request. If the userId is not available, the function returns null. Otherwise, it uses the userId to query the database and retrieve the corresponding profile using the db.profile.findUnique method. The retrieved profile is then returned by the function.