✨ feat(chat-welcome.tsx): add ChatWelcome component to display a welcome message in the chat
The ChatWelcome component is responsible for rendering a welcome message in the chat. It takes two props: `type` (either 'channel' or 'conversation') and `name` (the name of the channel or conversation).

If the `type` prop is set to 'channel', the component will display a circular icon with the '#' symbol using the Lucide React library. Otherwise, it will not display the icon.

The component also renders a heading with the channel or conversation name, and a paragraph with a description depending on the `type` prop. If `type` is 'channel', the description will indicate that it is the beginning of the channel. If `type` is 'conversation', the description will indicate that it is the start of the conversation with the specified name.